+++ DLMSProber/DLMSProber/src/GXDLMSClient.cpp
@@ -44,8 +44,8 @@
 
 
 CGXDLMSClient::CGXDLMSClient(bool UseLogicalNameReferencing,
-    int clientAddress,
-    int serverAddress,
+    unsigned long clientAddress,
+    unsigned long serverAddress,
     //Authentication type.
     DLMS_AUTHENTICATION authentication,
     //Password if authentication is used.
